Test and fix issues with gzipped output

+++ modules/nf-core/tiara/tiara/main.nf
@@ -12,10 +12,10 @@ process TIARA_TIARA {
     tuple val(meta), path(fasta)
 
     output:
-    tuple val(meta), path("classification_*.out.txt") , emit: classifications
-    tuple val(meta), path("log_*.out.txt")            , emit: log
-    tuple val(meta), path("*.fasta*")                 , emit: fasta, optional: true
-    path "versions.yml"                               , emit: versions
+    tuple val(meta), path("classification_*.out.{txt,txt.gz}")      , emit: classifications
+    tuple val(meta), path("log_*.out.{txt,txt.gz}")                 , emit: log
+    tuple val(meta), path("*.{fasta,fasta.gz}")                     , emit: fasta, optional: true
+    path "versions.yml"                                             , emit: versions
 
     when:
     task.ext.when == null || task.ext.when
@@ -30,7 +30,23 @@ process TIARA_TIARA {
         --threads ${task.cpus} \
         ${args}
 
-    mv ${prefix}.out.txt classification_${prefix}.out.txt
+    ## Rename classification files
+    if [ -f ${prefix}.out.txt ]; then
+        mv ${prefix}.out.txt classification_${prefix}.out.txt
+    else
+        mv ${prefix}.out.txt.gz classification_${prefix}.out.txt.gz
+    fi
+
+    ## fix gzip flag weirdness and ensure consistent .fasta filename output
+    ## check if fasta files are being output
+    if echo "${args}" | grep -q "tf\|to-fasta"; then
+        ## check if we've asked for gzip output, then rename files consistently
+        if echo "${args}" | grep -q "gz"; then
+            find . -name "*_${fasta}*" -exec sh -c 'file=`basename {}`; mv "\$file" "\${file%%.*}.fasta.gz"' \\;
+        else
+            find . -name "*_${fasta}*" -exec sh -c 'file=`basename {}`; mv "\$file" "\${file%%.*}.fasta"' \\;
+        fi
+    fi
 
     cat <<-END_VERSIONS > versions.yml
     "${task.process}":
